Tried to fix vcpkg error

+++ src/DSEngine/vcpkg-restore.bat
@@ -4,17 +4,26 @@ set _solutionDir=%2%
 
 if %_platform%==Win32 ( set _platform=x86 )
 
-if exist %_solutionDir%vcpkg ( 
-    if exist %_solutionDir%vcpkg\vcpkg.exe (
-        echo Found vcpkg
+WHERE vcpkg >nul 2>nul
+IF %ERRORLEVEL% NEQ 0 ( 
+    ECHO vcpkg not found in PATH, checking solution folder
+    if exist %_solutionDir%vcpkg ( 
+        if exist %_solutionDir%vcpkg\vcpkg.exe (
+            echo Found vcpkg
+        ) else (
+            %_solutionDir%vcpkg\bootstrap-vcpkg.bat
+        )
     ) else (
+        ECHO vcpkg not found
+        git clone https://github.com/Microsoft/vcpkg.git %_solutionDir%vcpkg
         %_solutionDir%vcpkg\bootstrap-vcpkg.bat
     )
+    ECHO Restoring package
+    %_solutionDir%vcpkg\vcpkg.exe integrate install
+    %_solutionDir%vcpkg\vcpkg.exe install boost:%_platform%-windows
 ) else (
-    ECHO vcpkg not found
-    git clone https://github.com/Microsoft/vcpkg.git %_solutionDir%vcpkg
-    %_solutionDir%vcpkg\bootstrap-vcpkg.bat
+    ECHO Using pre-installed vcpkg.
+    ECHO Restoring package
+    vcpkg integrate install
+    vcpkg install boost:%_platform%-windows
 )
-ECHO Restoring package
-%_solutionDir%vcpkg\vcpkg.exe integrate install
-%_solutionDir%vcpkg\vcpkg.exe install boost:%_platform%-windows
